+++ whisker_serial_order/extra.py
@@ -2,6 +2,8 @@
 # whisker_serial_order/extra.py
 
 """
+===============================================================================
+
     Copyright © 2016-2018 Rudolf Cardinal (rudolf@pobox.com).
 
     Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,6 +17,11 @@
     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
     See the License for the specific language governing permissions and
     limitations under the License.
+
+===============================================================================
+
+Additional functions.
+
 """
 
 import datetime
@@ -26,6 +33,13 @@ TimeType = Union[datetime.datetime, arrow.Arrow]
 
 def latency_s(t1: Optional[TimeType],
               t2: Optional[TimeType]) -> Optional[float]:
+    """
+    Calculates the latency in seconds between two datetime-type objects.
+
+    :param t1: start time
+    :param t2: end time
+    :return: time difference in seconds, or ``None`` if either were ``None``
+    """
     if t1 is None or t2 is None:
         return None
     delta = t2 - t1